fix(Maintainability) Address peer feedback (#2850303)
