+++ schemata.permissions.yml
@@ -1,2 +1,2 @@
-access data models:
+access schemata data models:
   title: 'Access the different data models'
