+++ schemata.services.yml
@@ -2,7 +2,7 @@ services:
   serializer.encoder.json_schema:
     class: Drupal\schemata\Encoder\JsonSchemaEncoder
     tags:
-      - { name: encoder, priority: 22, format: 'schema_json' }
+      - { name: encoder, priority: 10, format: 'schema_json' }
 
   # Assemble Schemas based on entity type definitions. As a standalone class,
   # the SchemaFactory can be pulled in to custom menu routes, drush commands,
