+++ schemata_json_schema/src/Annotation/TypeMapper.php
@@ -7,7 +7,7 @@ use Drupal\Component\Annotation\Plugin;
 /**
  * Defines a JSON Schema Type Mapper annotation object.
  *
- * Plugin Namespace: Plugin\json_schema\type_mapper
+ * Plugin Namespace: Plugin\schemata_json_schema\type_mapper
  *
  * @ingroup third_party
  *
