+++ schemata_json_schema/src/Plugin/Type/TypeMapperPluginManager.php
@@ -5,14 +5,17 @@ namespace Drupal\schemata_json_schema\Plugin\Type;
 use Drupal\Core\Plugin\DefaultPluginManager;
 use Drupal\Core\Extension\ModuleHandlerInterface;
 use Drupal\Component\Plugin\FallbackPluginManagerInterface;
+use Drupal\schemata_json_schema\Annotation\TypeMapper;
+use Drupal\schemata_json_schema\Plugin\schemata_json_schema\type_mapper\TypeMapperInterface;
+
 /**
  * Manages TypeMapper plugins.
  *
  * TypeMappers are used to adapt Drupal TypedData types to JSON Schema specs.
  *
  * @see \Drupal\schemata_json_schema\Annotation\TypeMapper
- * @see \Drupal\schemata_json_schema\Plugin\type_mapper\TypeMapperBase
- * @see \Drupal\schemata_json_schema\Plugin\type_mapper\TypeMapperInterface
+ * @see \Drupal\schemata_json_schema\Plugin\schemata_json_schema\type_mapper\TypeMapperBase
+ * @see \Drupal\schemata_json_schema\Plugin\schemata_json_schema\type_mapper\TypeMapperInterface
  * @see plugin_api
  */
 class TypeMapperPluginManager extends DefaultPluginManager implements FallbackPluginManagerInterface {
@@ -34,7 +37,7 @@ class TypeMapperPluginManager extends DefaultPluginManager implements FallbackPl
    *   The module handler to invoke the alter hook with.
    */
   public function __construct(\Traversable $namespaces, ModuleHandlerInterface $module_handler) {
-    parent::__construct('Plugin/type_mapper', $namespaces, $module_handler, 'Drupal\schemata_json_schema\Plugin\type_mapper\TypeMapperInterface', 'Drupal\schemata_json_schema\Annotation\TypeMapper');
+    parent::__construct('Plugin/schemata_json_schema/type_mapper', $namespaces, $module_handler, TypeMapperInterface::class, TypeMapper::class);
   }
 
   /**
