--- schemata_json_schema/src/Plugin/type_mapper/FilterFormatTypeMapper.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace Drupal\schemata_json_schema\Plugin\type_mapper;
-
-use Drupal\Core\TypedData\DataDefinitionInterface;
-
-/**
- * Converts Data Definition properties of filter_format type to JSON Schema.
- *
- * @TypeMapper(
- *  id = "filter_format"
- * )
- */
-class FilterFormatTypeMapper extends TypeMapperBase {
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getMappedValue(DataDefinitionInterface $property) {
-    $value = parent::getMappedValue($property);
-    $value['type'] = 'string';
-    return $value;
-  }
-
-}
